requirements.txt
Flask
Flask-Bcrypt
flask-login
Flask-SQLAlchemy
Flask-WTF

python-dotenv
